DRY up a bit, make logic super easy to read
